requirements.txt
django
djangorestframework
python-graphql-client
